# docs/api/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "The Swagger API documentation of the WordsApp",
        "version": "0.0.1"
    },
    "servers": [
        {
      "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "BasicAuth": []
        }
    ],
...
    "paths": {
...
        "/api/groups": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "Get a list of account's groups",
                "operationId": "get",
                "responses": {
                    "200": {
                        "description": "Groups found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GroupDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Group"
                ],
                "summary": "Update group",
                "operationId": "update_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGroupDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Return updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Group not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Save new group",
                "operationId": "save",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveGroupDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Group saved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Group already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            }
        },
...
        "/api/words/groups/{groupId}": {
            "get": {
                "tags": [
                    "Word"
                ],
                "summary": "Get all words paginated from a group",
                "operationId": "get_2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Words found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WordDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Word"
                ],
                "summary": "Save new word to group",
                "operationId": "save_1",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveWordRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Word saved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WordDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Word already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WordDto"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            }
        },
...
        "/api/words/groups/{groupId}/{wordId}": {
            "delete": {
                "tags": [
                    "Word"
                ],
                "summary": "Delete word from group",
                "operationId": "delete_1",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "wordId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Word deleted or ignored"
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            }
        },
        "/api/groups/{id}": {
            "delete": {
                "tags": [
                    "Group"
                ],
                "summary": "Delete group",
                "description": "Delete a group even if it doesn't exist",
                "operationId": "delete_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group deleted or ignored"
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "WordDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "languageId": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "definition": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "translations": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "transcriptions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "created",
                    "id",
                    "languageId",
                    "title"
                ]
            },
            "UpdateGroupDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "description",
                    "id",
                    "title"
                ]
            },
            "GroupDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "created": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    },
                    "ownerId": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    }
                },
                "required": [
                    "created",
                    "description",
                    "id",
                    "ownerId",
                    "title"
                ]
            },
...
            "SaveWordRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "languageId": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "definition": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "transcriptions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "translations": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "languageId",
                    "title"
                ]
            },
            "SaveGroupDto": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "description",
                    "title"
                ]
            },
...
        },
        "securitySchemes": {
            "BasicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}